chore: rename file to match the hooks purpose
