+++ src/modules/meetings/ui/components/meetings-search-filter.tsx
@@ -1,7 +1,7 @@
 import { SearchIcon } from 'lucide-react';
 
 import { Input } from '@/components/ui/input';
-import { useMeetingsFilters } from '@/modules/meetings/hooks/use-agents-filters';
+import { useMeetingsFilters } from '@/modules/meetings/hooks/use-meetings-filters';
 
 export const MeetingsSearchFilter = () => {
   const [filters, setFilters] = useMeetingsFilters();
