+++ src/modules/meetings/ui/views/meetings-view.tsx
@@ -8,7 +8,7 @@ import { EmptyState } from '@/components/empty-state';
 import { ErrorState } from '@/components/error-state';
 import { LoadingState } from '@/components/loading-state';
 import { DataTable } from '@/components/ui/data-table';
-import { useMeetingsFilters } from '@/modules/meetings/hooks/use-agents-filters';
+import { useMeetingsFilters } from '@/modules/meetings/hooks/use-meetings-filters';
 import { meetingsColumns } from '@/modules/meetings/ui/components/meetings-columns';
 import { useTRPC } from '@/utils/trpc';
